+++ packages/llamaindex/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/llamaindex
 
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [ab7ccef]
+- Updated dependencies [257224b]
+- Updated dependencies [0463011]
+- Updated dependencies [d306260]
+  - ai@5.0.0-alpha.3
+
 ## 1.0.0-alpha.2
 
 ### Patch Changes